Logging Updated
